requirements.txt
 streamlit
scikit-learn
opencv-python-headless
numpy
pillow
joblib